make inline tests pass

+++ src/subframe.rs
@@ -491,16 +491,17 @@ fn predict_fixed<Sample: sample::WideSample>
 #[test]
 fn verify_predict_fixed() {
     // The following data is from an actual FLAC stream and has been verified
-    // against the reference decoder.
+    // against the reference decoder. The data is from a 16-bit stream, so the
+    // wide sample type is `i32`.
     let mut buffer = [-729, -722, -667, -19, -16,  17, -23, -7,
                         16,  -16,   -5,   3,  -8, -13, -15, -1];
-    assert!(predict_fixed::<i16>(3, &mut buffer).is_ok());
+    assert!(predict_fixed::<i32>(3, &mut buffer).is_ok());
     assert_eq!(&buffer, &[-729, -722, -667, -583, -486, -359, -225, -91,
                             59,  209,  354,  497,  630,  740,  812, 845]);
 
     // The following data causes overflow when not handled with care.
     let mut buffer = [21877, 27482, -6513];
-    assert!(predict_fixed::<i16>(2, &mut buffer).is_ok());
+    assert!(predict_fixed::<i32>(2, &mut buffer).is_ok());
     assert_eq!(&buffer, &[21877, 27482, 26574]);
 }
 
@@ -575,18 +576,19 @@ fn predict_lpc<Sample: sample::WideSample>
 #[test]
 fn verify_predict_lpc() {
     // The following data is from an actual FLAC stream and has been verified
-    // against the reference decoder.
+    // against the reference decoder. The data is from a 16-bit stream, so the
+    // wide sample type is `i32`.
     let coefficients = [-75, 166,  121, -269, -75, -399, 1042];
     let mut buffer = [-796, -547, -285,  -32, 199,  443,  670, -2,
                        -23,   14,    6,    3,  -4,   12,   -2, 10];
-    assert!(predict_lpc::<i16>(&coefficients, 9, &mut buffer).is_ok());
+    assert!(predict_lpc::<i32>(&coefficients, 9, &mut buffer).is_ok());
     assert_eq!(&buffer, &[-796, -547, -285,  -32,  199,  443,  670,  875,
                           1046, 1208, 1343, 1454, 1541, 1616, 1663, 1701]);
 
     // The following data causes an overflow when not handled with care.
     let coefficients = [119, -255, 555, -836, 879, -1199, 1757];
     let mut buffer = [-21363, -21951, -22649, -24364, -27297, -26870, -30017, 3157];
-    assert!(predict_lpc::<i16>(&coefficients, 10, &mut buffer).is_ok());
+    assert!(predict_lpc::<i32>(&coefficients, 10, &mut buffer).is_ok());
     assert_eq!(&buffer, &[-21363, -21951, -22649, -24364, -27297, -26870, -30017, -29718]);
 }
